mea culpa
If I'm going to provide the user with the power to configure the basic type of entities within a game, I should make it easier for the user, and the developer, to handle problems with these mod files vs. the way the in game objects work.

Separate out the entity base configuration into a `config-game-objects` module that will handle just the basic configuration validation, loading of the validation files (right now from disk), and whatever other crud around potential user errors we don't want mixed into potential developer errors in the code.