experiment with i18next formatters
separating color, as a style, from the content, is the effort, and the mechanism to remove it from being mixed in with the translation strings led to trying out the i18next formatters. It feels less gross than what I was doing before, but still feels a bit icky. Maybe a post processor is needed. Also needed is a more generic colorization function that can be toggled off for no color display without hurting anything.