intermediate checkin
want to adhere to a key:value pattern for storing data. We'll see how long this works, but it let's me pull out the annoying file system stuff that keeps creeping in and hides it in the new io module. The testing-grounds is in an incomplete state, but whatever, it's the momentary chaos central.