naming no longer part of game-objects
naming is weird and confusing and making it play by the other entity rules felt like it being destined to break. Naming might get a bit funky if it becomes popular, needs to prevent dupes, needs to be more procedural, etc... but for now it is fine for it to live in simulation.