testing-grounds formalized as launcher
testing-grounds is just a launcher, and programs run by it should be self-contained, yielding back to the launcher only when done. This removes the obtuse reliance I had at the top level of programs yielding a function to be run by the main launcher to yield control back to the sub-program. If sub-programs have thet potential to be run-looped, they should provide their own run loop and control flow outside of the launcher.