delete some strength code
Originally had the idea of plug and play functions for game rules, like what could affect strength and how. I still like that idea, but it's way too soon. The injected/injectable functions work, but they make me not want to work in the strength code right now, and the idea was just not fully planned out. It's easier to work in the code if things don't have a lot of indirection and I just write the code needed.